fix: refine menu bar pace display (#169) (thanks @PhilETaylor)
